+++ frontend/src/routes/Applications/ApplicationDetails/ApplicationTopology/model/computeStatuses.data.js
